backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
django-filter
gunicorn==20.0.4
psycopg2-binary==2.8.6
djangorestframework-simplejwt
djoser==2.1.0
pytz==2021.1
pillow
drf-extra-fields
python-dotenv==0.18.0